reset des interruptions sur décision du handler d'interrupt coté userspace:
s'il répond false, on ferme et on réouvre /dev/interrupts, ce qui a pour effet
de prévenir le module qu'il peut repartir d'un état initial et faire comme si aucune erreur ne s'était produite.

Cela implique que la partie userspace a stoppé la cause de l'erreur en premier lieu (ex: stop sequence)

+ augmentation taille jumptable d'interruptions a 256 et utilisation uint8_t au lieu d'int8_t pour permettre d'utiliser 0xFF comme marqueur d'erreur.

+++ kernelspace/gpio_irq.c
@@ -3,7 +3,7 @@
 #include "../common/interrupt_codes.h"
 
 //macros to declare associations between GPIO and interrupt code in a readable list
-#define _gpio(code, gpio) {#code, code, gpio, -EINVAL}
+#define _gpio(code, gpio) {#code, (uint8_t)code, gpio, -EINVAL}
 #define _gpio_list_end {NULL, -EINVAL, -EINVAL, -EINVAL}
 
 //associations between GPIO and interrupt codes
